requirements.txt
alabaster==0.7.12; python_version >= "3.7"
babel==2.11.0; python_version >= "3.7"
beautifulsoup4==4.11.1; python_full_version >= "3.6.0" and python_version >= "3.7"
boltons==21.0.0; python_version >= "3.7" and python_version < "4.0"
bracex==2.3.post1; python_version >= "3.7" and python_version < "4.0"
certifi==2022.12.7; python_version >= "3.7" and python_version < "4"
charset-normalizer==2.1.1; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.0"
click-extra==3.5.0; python_version >= "3.7" and python_version < "4.0"
click-log==0.4.0; python_version >= "3.7" and python_version < "4.0"
click-shell==2.1
click==8.1.3; python_version >= "3.7"
cloup==2.0.0.post1; python_version >= "3.7" and python_version < "4.0"
colorama==0.4.6; python_version >= "3.7" and python_full_version < "3.0.0" and platform_system == "Windows" and sys_platform == "win32" or platform_system == "Windows" and python_version >= "3.7" and python_full_version >= "3.7.0" and sys_platform == "win32"
colorlog==6.7.0; python_version >= "3.6"
commentjson==0.9.0; python_version >= "3.7" and python_version < "4.0"
docutils==0.17.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
furo==2022.12.7; python_version >= "3.7"
fuzzywuzzy==0.18.0
idna==3.4; python_version >= "3.7" and python_version < "4"
imagesize==1.4.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
jinja2==3.1.2; python_version >= "3.7"
lark-parser==0.7.8; python_version >= "3.7" and python_version < "4.0"
levenshtein==0.20.8; python_version >= "3.6"
lxml==4.9.2; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
markupsafe==2.1.1; python_version >= "3.7"
mergedeep==1.3.4; python_version >= "3.7" and python_version < "4.0"
numpy==1.23.5; python_version >= "3.8"
packaging==22.0; python_version >= "3.7" and python_version < "4.0"
pallets-sphinx-themes==2.0.2; python_version >= "3.7" and python_version < "4.0"
pygments-ansi-color==0.1.0; python_version >= "3.7" and python_version < "4.0"
pygments==2.13.0; python_version >= "3.7" and python_version < "4.0"
python-levenshtein==0.20.8; python_version >= "3.6"
pytz==2022.6; python_version >= "3.7"
pyyaml==6.0; python_version >= "3.6"
rapidfuzz==2.13.6; python_version >= "3.7"
regex==2022.10.31; python_version >= "3.7" and python_version < "4.0"
requests==2.28.1; python_version >= "3.7" and python_version < "4"
snowballstemmer==2.2.0; python_version >= "3.7"
soupsieve==2.3.2.post1; python_full_version >= "3.6.0" and python_version >= "3.7"
sphinx-basic-ng==1.0.0b1; python_version >= "3.7"
sphinx==5.3.0; python_version >= "3.6"
sphinxcontrib-applehelp==1.0.2; python_version >= "3.7"
sphinxcontrib-devhelp==1.0.2; python_version >= "3.7"
sphinxcontrib-htmlhelp==2.0.0; python_version >= "3.7"
sphinxcontrib-jsmath==1.0.1; python_version >= "3.7"
sphinxcontrib-qthelp==1.0.3; python_version >= "3.7"
sphinxcontrib-serializinghtml==1.1.5; python_version >= "3.7"
tabulate==0.9.0; python_version >= "3.7"
tomli==2.0.1; python_version >= "3.7" and python_version < "3.11"
urllib3==1.26.13; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4" or python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.0"
wcmatch==8.4.1; python_version >= "3.7" and python_version < "4.0"
wcwidth==0.2.5; python_version >= "3.7" and python_version < "4.0"
xmltodict==0.13.0; python_version >= "3.7" and python_version < "4.0"
